+++ Changelog.md
@@ -1,3 +1,7 @@
+--------------------------Changelog for v1.10-beta2 of SecurityCraft--------------------------
+
+- Fix: Jade integration
+
 --------------------------Changelog for v1.10-beta1 of SecurityCraft--------------------------
 
 - API: Merged IPasscodeProtected#loadSaltKey and IPasscodeProtected#loadPasscode into a single method